ref(rlm): Extract finalization syntax validation
Separate the SUBMIT grammar from DSPy compatibility code without changing\nrepair behavior. Cover accepted expressions and rejected exploration syntax.
